+++ src/spatialdata/_core/query/relational_query.py
@@ -19,6 +19,8 @@ from spatial_image import SpatialImage
 from spatialdata._core.spatialdata import SpatialData
 from spatialdata._utils import _inplace_fix_subset_categorical_obs
 from spatialdata.models import (
+    Image2DModel,
+    Image3DModel,
     Labels2DModel,
     Labels3DModel,
     PointsModel,
@@ -462,17 +464,50 @@ class MatchTypes(Enum):
     no = "no"
 
 
-def join_sdata_spatialelement_table(
-    sdata: SpatialData,
-    spatial_element_name: str | list[str],
-    table_name: str,
-    how: str = "left",
+def _create_sdata_elements_dict_for_join(
+    sdata: SpatialData, spatial_element_name: str | list[str]
+) -> dict[str, dict[str, Any]]:
+    elements_dict: dict[str, dict[str, Any]] = defaultdict(lambda: defaultdict(dict))
+    for name in spatial_element_name:
+        element_type, _, element = sdata._find_element(name)
+        elements_dict[element_type][name] = element
+    return elements_dict
+
+
+def _validate_element_types_for_join(
+    sdata: SpatialData | None,
+    spatial_element_names: list[str],
+    spatial_elements: list[SpatialElement] | None,
+    table: AnnData | None,
+) -> None:
+    if sdata is not None:
+        elements_to_check = []
+        for name in spatial_element_names:
+            elements_to_check.append(sdata[name])
+    else:
+        assert spatial_elements is not None
+        elements_to_check = spatial_elements
+
+    for element in elements_to_check:
+        model = get_model(element)
+        if model in [Image2DModel, Image3DModel, TableModel]:
+            raise ValueError(f"Element type `{model}` not supported for join operation.")
+
+
+def join_spatialelement_table(
+    sdata: SpatialData | None = None,
+    spatial_element_names: str | list[str] | None = None,
+    spatial_elements: SpatialElement | list[SpatialElement] | None = None,
+    table_name: str | None = None,
+    table: AnnData | None = None,
+    how: Literal["left", "left_exclusive", "inner", "right", "right_exclusive"] = "left",
     match_rows: Literal["no", "left", "right"] = "no",
 ) -> tuple[dict[str, Any], AnnData]:
-    """Join SpatialElement(s) and table together in SQL like manner.
+    """
+    Join SpatialElement(s) and table together in SQL like manner.
 
     The function allows the user to perform SQL like joins of SpatialElements and a table. The elements are not
-    returned together in one dataframe like structure, but instead filtered elements are returned. To determine matches,
+    returned together in one dataframe-like structure, but instead filtered elements are returned. To determine matches,
     for the SpatialElement the index is used and for the table the region key column and instance key column. The
     elements are not overwritten in the `SpatialData` object.
 
@@ -490,11 +525,19 @@ def join_sdata_spatialelement_table(
     Parameters
     ----------
     sdata
-        The SpatialData object containing the tables and spatial elements.
-    spatial_element_name
-        The name(s) of the spatial elements to be joined with the table.
+        SpatialData object containing all the elements and tables. This parameter can be `None`; in such case the both
+        the names and values for the elements and the table must be provided.
+    spatial_element_names
+        Required. The name(s) of the spatial elements to be joined with the table. If a list of names, and if sdata is
+         `None`, the indices must match with the list of SpatialElements passed on by the argument elements.
+    spatial_elements
+        This parameter should be speficied exactly when `sdata` is `None`. The SpatialElement(s) to be joined with the
+        table. In case of a list of SpatialElements the indices must match exactly with the indices in the list of
+        `spatial_element_name`.
     table_name
-        The name of the table to join with the spatial elements.
+        The name of the table to join with the spatial elements. Optional, `table` can be provided instead.
+    table
+        The table to join with the spatial elements. When `sdata` is not `None`, `table_name` can be used instead.
     how
         The type of SQL like join to perform, default is ``'left'``. Options are ``'left'``, ``'left_exclusive'``,
         ``'inner'``, ``'right'`` and ``'right_exclusive'``.
@@ -508,35 +551,61 @@ def join_sdata_spatialelement_table(
 
     Raises
     ------
-    AssertionError
-        If no table with the given table_name exists in the SpatialData object.
+    ValueError
+        If `spatial_element_names` is not provided.
+    ValueError
+        If sdata is `None` but `spatial_elements` is not `None`; if `sdata` is not `None`, but `spatial_elements` is
+        `None`.
+    ValueError
+        If `table_name` is provided but not present in the `SpatialData` object, or if `table_name` is provided but
+        `sdata` is `None`.
+    ValueError
+        If not exactly one of `table_name` and `table` is provided.
+    ValueError
+        If no valid elements are provided for the join operation.
     ValueError
         If the provided join type is not supported.
+    ValueError
+        If an incorrect value is given for `match_rows`.
     """
-    assert sdata.tables.get(table_name), f"No table with `{table_name}` exists in the SpatialData object."
-    table = sdata.tables[table_name]
-    if isinstance(spatial_element_name, str):
-        spatial_element_name = [spatial_element_name]
+    if spatial_element_names is None:
+        raise ValueError("`spatial_element_names` must be provided.")
+    if sdata is None and (spatial_elements is None or table is None):
+        raise ValueError("If `sdata` is not provided, both `spatial_elements` and `table` must be provided.")
+    if sdata is not None and (spatial_elements is not None):
+        raise ValueError(
+            "If `sdata` is provided, `spatial_elements` must not be provided; use `spatial_elements_name` instead."
+        )
+    if table is None and table_name is None or table is not None and table_name is not None:
+        raise ValueError("Exactly one of `table_name` and `table` must be provided.")
+    if sdata is not None and table_name is not None:
+        if table_name not in sdata.tables:
+            raise ValueError(f"No table with name `{table_name}` found in the SpatialData object.")
+        table = sdata[table_name]
+    spatial_element_names = (
+        spatial_element_names if isinstance(spatial_element_names, list) else [spatial_element_names]
+    )
+    spatial_elements = spatial_elements if isinstance(spatial_elements, list) else [spatial_elements]
+    _validate_element_types_for_join(sdata, spatial_element_names, spatial_elements, table)
 
-    elements_dict: dict[str, dict[str, Any]] = defaultdict(lambda: defaultdict(dict))
-    for name in spatial_element_name:
-        if name in sdata.tables:
-            warnings.warn(
-                f"Tables: `{', '.join(elements_dict['tables'].keys())}` given in spatial_element_names cannot be "
-                f"joined with a table using this function.",
-                UserWarning,
-                stacklevel=2,
-            )
-        elif name in sdata.images:
-            warnings.warn(
-                f"Images: `{', '.join(elements_dict['images'].keys())}` cannot be joined with a table",
-                UserWarning,
-                stacklevel=2,
-            )
-        else:
-            element_type, _, element = sdata._find_element(name)
-            elements_dict[element_type][name] = element
+    elements_dict: dict[str, dict[str, Any]]
+    if sdata is not None:
+        elements_dict = _create_sdata_elements_dict_for_join(sdata, spatial_element_names)
+    else:
+        derived_sdata = SpatialData.from_elements_dict(dict(zip(spatial_element_names, spatial_elements)))
+        element_types = ["labels", "shapes", "points"]
+        elements_dict = defaultdict(lambda: defaultdict(dict))
+        for element_type in element_types:
+            for name, element in getattr(derived_sdata, element_type).items():
+                elements_dict[element_type][name] = element
 
+    elements_dict_joined, table = _call_join(elements_dict, table, how, match_rows)
+    return elements_dict_joined, table
+
+
+def _call_join(
+    elements_dict: dict[str, dict[str, Any]], table: AnnData, how: str, match_rows: Literal["no", "left", "right"]
+) -> tuple[dict[str, Any], AnnData]:
     assert any(key in elements_dict for key in ["labels", "shapes", "points"]), (
         "No valid element to join in spatial_element_name. Must provide at least one of either `labels`, `points` or "
         "`shapes`."
@@ -611,7 +680,9 @@ def match_element_to_table(
     -------
     A tuple containing the joined elements as a dictionary and the joined table as an AnnData object.
     """
-    element_dict, table = join_sdata_spatialelement_table(sdata, element_name, table_name, "right", match_rows="right")
+    element_dict, table = join_spatialelement_table(
+        sdata=sdata, spatial_element_names=element_name, table_name=table_name, how="right", match_rows="right"
+    )
     return element_dict, table