+++ CHANGELOG.md
@@ -8,7 +8,17 @@ and this project adheres to [Semantic Versioning][].
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [0.0.x] - tbd
+## [0.1.1] - xxxx-xx-xx
+
+### Added
+
+-   Added method `update_annotated_regions_metadata() which updates the `region`value automatically from the `region_key` columns
+
+### Changed
+
+-   Renamed `join_sdata_spatialelement_table` to `join_spatialelement_table`, and made it work also without `SpatialData` objects.
+
+## [0.1.0] - 2024-03-24
 
 ### Added